requirements.txt
asgiref==3.5.0
Django==4.0.2
django-cors-headers==3.11.0
django-environ==0.8.1
django-tinymce4-lite==1.8.0
djangorestframework==3.13.1
gunicorn==20.1.0
jsmin==3.0.1
psycopg2-binary==2.9.3
pygame==2.1.0
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
